Move Decoder-specific transforms to separate package
