+++ ingestion-beam/src/main/java/com/mozilla/telemetry/decoder/ParseUri.java
@@ -2,8 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package com.mozilla.telemetry.transforms;
+package com.mozilla.telemetry.decoder;
 
+import com.mozilla.telemetry.transforms.FailureMessage;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.beam.sdk.io.gcp.pubsub.PubsubMessage;
