+++ ingestion-beam/src/main/java/com/mozilla/telemetry/decoder/ParseUserAgent.java
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package com.mozilla.telemetry.transforms;
+package com.mozilla.telemetry.decoder;
 
 import com.google.common.collect.Sets;
 import java.io.Serializable;
