+++ ingestion-beam/src/test/java/com/mozilla/telemetry/DecoderTest.java
@@ -7,10 +7,10 @@ package com.mozilla.telemetry;
 import com.mozilla.telemetry.options.InputFileFormat;
 import com.mozilla.telemetry.options.OutputFileFormat;
 import com.mozilla.telemetry.transforms.DecodePubsubMessages;
-import com.mozilla.telemetry.transforms.GeoCityLookup;
-import com.mozilla.telemetry.transforms.GzipDecompress;
-import com.mozilla.telemetry.transforms.ParseUri;
-import com.mozilla.telemetry.transforms.ParseUserAgent;
+import com.mozilla.telemetry.decoder.GeoCityLookup;
+import com.mozilla.telemetry.decoder.GzipDecompress;
+import com.mozilla.telemetry.decoder.ParseUri;
+import com.mozilla.telemetry.decoder.ParseUserAgent;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.beam.sdk.testing.PAssert;
